backend/requirements.txt
# ========================================
# Dependencias del Backend ML - Litper Logística
# ========================================

# Framework Web
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Base de Datos
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1

# Procesamiento de Datos
pandas==2.2.0
numpy==1.26.4
openpyxl==3.1.2

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.3.2

# IA - Anthropic Claude
anthropic==0.18.1

# Tareas Programadas
apscheduler==3.10.4

# Validación y Serialización
pydantic==2.6.1
pydantic-settings==2.1.0

# Utilidades
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.26.0

# Testing (opcional)
pytest==8.0.0
pytest-asyncio==0.23.5

# Herramientas de Desarrollo (opcional)
black==24.1.1
isort==5.13.2
flake8==7.0.0
